requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
streamlit>=1.28.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
scipy>=1.10.0
requests>=2.31.0
tqdm>=4.65.0

